requirements.txt
asgiref==3.11.0
bleach==6.3.0
Django==6.0
django-appconf==1.2.0
django-extensions==4.1
django-ranged-response==0.2.0
django-robots==6.1
django-simple-captcha==0.6.3
django-summernote==0.8.20.0
django-taggit==6.1.0
django_compressor==4.6.0
pillow==12.0.0
python-dotenv==1.2.1
rcssmin==1.2.2
rjsmin==1.2.5
sqlparse==0.5.4
tzdata==2025.3
webencodings==0.5.1
psycopg2-binary